Refactor common utils out to enrollment-testutils
