+++ src/components/dashboard/main-content/course-enrollments/tests/CourseEnrollments.test.jsx
@@ -3,65 +3,50 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 import { Provider } from 'react-redux';
-import thunk from 'redux-thunk';
+
+// test deps
 import configureMockStore from 'redux-mock-store';
 
 // requirements
 import { AppContext } from '@edx/frontend-platform/react';
 import { getAuthenticatedUser } from '@edx/frontend-platform/auth';
+import {
+  createMockStore,
+  createCompletedCourseRun,
+  defaultInitialEnrollmentProps,
+} from './enrollment-testutils';
 
 // component to test
 import { CourseEnrollments } from '../CourseEnrollments';
 
+// TODO: mock auth, we should consider using a authprovider in the test
 jest.mock('@edx/frontend-platform/auth');
 getAuthenticatedUser.mockReturnValue({ username: 'test-username' });
 
-const mockStore = configureMockStore([thunk]);
 
 const genericMockFn = () => jest.fn();
+const store = createMockStore(configureMockStore);
 
-const completedCourseRun = {
-  courseRunId: 'course-v1:edX+DemoX+Demo_Course',
-  courseRunStatus: 'completed',
-  linkToCourse: 'https://edx.org/',
-  title: 'edX Demonstration Course',
-  notifications: [],
-  startDate: '2017-02-05T05:00:00Z',
-  endDate: '2018-08-18T05:00:00Z',
-  hasEmailsEnabled: true,
-  markedDone: false,
+const enterpriseConfig = {
+  uuid: 'test-enterprise-uuid',
 };
-
+const completedCourseRun = createCompletedCourseRun();
 const inProgCourseRun = { ...completedCourseRun, courseRunStatus: 'in_progress' };
-const store = mockStore({
-  emailSettings: {
-    loading: false,
-    error: null,
-    data: null,
-  },
-});
 
+const defaultInitialProps = defaultInitialEnrollmentProps({ genericMockFn });
+
+// TODO not sure why and if we need the sidebarComponent here
 const initialProps = {
+  ...defaultInitialProps,
   courseRuns: {
     in_progress: [inProgCourseRun],
     upcoming: [],
     completed: [completedCourseRun],
   },
-  isLoading: false,
-  error: null,
   sidebarComponent: <div className="sidebar-example" />,
-  fetchCourseEnrollments: genericMockFn(),
-  clearCourseEnrollments: genericMockFn(),
-  isMarkCourseCompleteSuccess: false,
-  modifyIsMarkCourseCompleteSuccess: genericMockFn(),
-  isUnarchiveCourseSuccess: false,
-  modifyIsUnarchiveCourseSuccess: genericMockFn(),
 };
 
 test('loads enrollments component', async () => {
-  const enterpriseConfig = {
-    uuid: 'test-enterprise-uuid',
-  };
 
   render(
     <Provider store={store}>
